Added new type of java constant generation
New setting javaFactory: enum/class allows for choosing between
an enum or a class with public static final members as generated
java constant file. Default is the old enum generation.

The reason behind this is that java does not allow for methods
larger than 64KB and the initilization of an enum becomes one
method. So if you have an enum with a lot of constants it is easy
to get a generated initialization larger than 64KB.